attackPoint changed from double to int in Card added Constructor parameter in Deck added add cards to a deck function in PlayerController

+++ beckend_gruppe_q/src/main/java/de/unidue/beckend_gruppe_q/model/Card.java
@@ -17,13 +17,13 @@ public class Card {
     private Long id;
     private String cardName;
     private Rarity cardRarity;
-    private double attackPoints;
+    private int attackPoints;
     private int defensePoints;
     private String description;
     private String image;
 
 
-    public Card( String cardName, Rarity cardRarity, double attackPoints, int defensePoints, String description, String image) {
+    public Card( String cardName, Rarity cardRarity, int attackPoints, int defensePoints, String description, String image) {
         this.cardName = cardName;
         this.cardRarity = cardRarity;
         this.attackPoints = attackPoints;
